macro-nutrient/requirements.txt
Flask>=2.0.0
Flask-JWT-Extended>=4.0.0
firebase-admin
google-cloud-firestore
google-cloud-storage
gunicorn
tensorflow>=2.18.0
protobuf<=3.20.3
Pillow
flask-cors>=6.0.0